pyproject.toml
[project]
name = "ma-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.96.0",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "mcp>=1.27.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "uvicorn>=0.45.0",
]
